- Core: im_project_nuke:   Now dealing with the situation that im_biz_object_groups   may not yet have installed

+++ tcl/intranet-project-procs.tcl
@@ -2957,9 +2957,12 @@ ad_proc im_project_nuke {
 		where program_id = :project_id"
 	}
 
-	db_dml delete_biz_object_groups "
+	if {[im_table_exists biz_object_groups]} {
+	    db_dml delete_biz_object_groups "
 		delete from biz_object_groups 
 		where biz_object_id = :project_id"
+	}
+
 	db_dml delete_projects "
 		delete from im_projects 
 		where project_id = :project_id"
